Move inbox state methods into inbox_state mod
