+++ bindings_wasm/src/inbox_state.rs
@@ -1,6 +1,8 @@
-use wasm_bindgen::prelude::wasm_bindgen;
+use wasm_bindgen::{prelude::wasm_bindgen, JsError};
 use xmtp_cryptography::signature::ed25519_public_key_to_address;
 use xmtp_id::associations::{AssociationState, MemberIdentifier};
+
+use crate::mls_client::WasmClient;
 
 #[wasm_bindgen(getter_with_clone)]
 #[derive(Clone)]
@@ -64,5 +66,39 @@ impl From<AssociationState> for WasmInboxState {
         .collect(),
       account_addresses: state.account_addresses(),
     }
+  }
+}
+
+#[wasm_bindgen]
+impl WasmClient {
+  /**
+   * Get the client's inbox state.
+   *
+   * If `refresh_from_network` is true, the client will go to the network first to refresh the state.
+   * Otherwise, the state will be read from the local database.
+   */
+  #[wasm_bindgen(js_name = inboxState)]
+  pub async fn inbox_state(&self, refresh_from_network: bool) -> Result<WasmInboxState, JsError> {
+    let state = self
+      .inner_client()
+      .inbox_state(refresh_from_network)
+      .await
+      .map_err(|e| JsError::new(format!("{}", e).as_str()))?;
+    Ok(state.into())
+  }
+
+  #[wasm_bindgen(js_name = getLatestInboxState)]
+  pub async fn get_latest_inbox_state(&self, inbox_id: String) -> Result<WasmInboxState, JsError> {
+    let conn = self
+      .inner_client()
+      .store()
+      .conn()
+      .map_err(|e| JsError::new(format!("{}", e).as_str()))?;
+    let state = self
+      .inner_client()
+      .get_latest_association_state(&conn, &inbox_id)
+      .await
+      .map_err(|e| JsError::new(format!("{}", e).as_str()))?;
+    Ok(state.into())
   }
 }
